requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.3.0
psutil>=5.8.0
winotify>=1.0.4; sys_platform == "win32"
pysmb>=1.2.7
pyautogui>=0.9.53
